Petistaan

+++ Petistaan/src/main/java/com/abhishekvermaa10/service/impl/OwnerServiceImpl.java
@@ -1,6 +1,7 @@
 package com.abhishekvermaa10.service.impl;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -17,6 +18,7 @@ import com.abhishekvermaa10.service.OwnerService;
  */
 @Service
 public class OwnerServiceImpl implements OwnerService {
+	
 	private final OwnerRepository ownerRepository;
 	private final String ownerAlreadyExists;
 	private final String ownerNotFound;
@@ -46,20 +48,31 @@ public class OwnerServiceImpl implements OwnerService {
 
 	@Override
 	public void updatePetDetails(int ownerId, String petName) throws OwnerNotFoundException {
-		ownerRepository.findById(ownerId)
+		OwnerDTO ownerDTO = ownerRepository.findById(ownerId)
 				.orElseThrow(() -> new OwnerNotFoundException(String.format(ownerNotFound, ownerId)));
-		ownerRepository.updatePetDetails(ownerId, petName);
+		ownerDTO.getPetDTO().setName(petName);
+		ownerRepository.save(ownerDTO);
 	}
 
 	@Override
 	public void deleteOwner(int ownerId) throws OwnerNotFoundException {
-		ownerRepository.findById(ownerId)
-				.orElseThrow(() -> new OwnerNotFoundException(String.format(ownerNotFound, ownerId)));
-		ownerRepository.deleteById(ownerId);
+		Optional<OwnerDTO> optionalOwner = ownerRepository.findById(ownerId);
+		if (optionalOwner.isEmpty()) {
+			throw new OwnerNotFoundException(String.format(ownerNotFound, ownerId));
+		} else {
+			ownerRepository.deleteById(ownerId);
+		}
 	}
 
 	@Override
 	public List<OwnerDTO> findAllOwners() {
 		return ownerRepository.findAll();
 	}
+	
+	@Override
+	public List<Object[]> findIdAndFirstNameAndLastNameAndPetNameOfPaginatedOwners(int pageNumber,
+			int numberOfRecordsPerPage) {
+		return ownerRepository.findIdAndFirstNameAndLastNameAndPetName(pageNumber, numberOfRecordsPerPage);
+	}
+	
 }
